admin/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
PyJWT[crypto]
jinja2
python-multipart
